Fix calls to jsonp. The problem is that jsonp does not pass on status codes. Because this is the case calls to our server that specify jsonp get wrapped with extra information that indicates whether or not the call was successful, along with the status code.
Started fixing the parts of the code that call $http.jsonp() to instead use
the new service that wraps.

Changed the notification directive to reduce the display time to 3 seconds.

Fixed the EventController to initialize $scope.mesage. The message variable
was not being set which was causing a lot of errors in the console due to
an interaction with angular-strap.