ISSUE#62 In some cases angular is throwing an error when calling a function assigned to the $rootScope. This occurs when the call is performed in a partial page. It is unclear why some pages are working and some aren't. The particular call that is affected is $rootScope.toDateString().
The fix for this issue was to create some variables in the $scope that
represent the date and to call $rootScope.toDateString() in the controller.

As a side effect of this change I also renamed and moved the DataSearchController.
The new name is DataDirReportController, and I moved it into the DataGroupsControllers.js
file. I also changed the name of the partial and the route to reflect that it is
working on datadirs.