ISSUE#64
Converted toDateString() to a filter and moved it out of $rootScope. Updated
all affected partials and javascript code. One odd thing - filters or the
$rootScope function are being called multiple times by angular. Sometimes
with data, and sometimes without data.

The "fix" for this is to make sure that the parameter in the function
actually has data before attempting to do anything with it.