ISSUE#121
The call to create the /tmp/uploads directory was taking place after
we attempted to create a temporary directory in /tmp/uploads. Thus the
temporary directory creation would fail if /tmp/uploads didn't exist.

I moved the call to create /tmp/uploads to before the call to create
the temporary directory.