ISSUE#555 ISSUE#542
Aligned counts to the right and put them in a badge to make them easier
to see. Moved icons to the left so that they align. Having them to the
right of the text makes alignment more fiddly.

Fixed updating the counts for notes and reviews when they change.

Code cleanup.