Objects in memory are being modified. That means logout/login keeps these objects around and thus showed incorrect favorite state. This is a bad design, however I created a fix within the design by resetting all favorite states.
Changed selection of template so that you have to click on a box. UI wise it doesn't look great so we'll need to figure out what to do with the look. However functionality wise it is much friendlier than clicking on setup or prefill and then having that template selected such that clicking ok goes to create process even though that isn't what you wanted.

Removed the ok button, and just have a dismiss button since the template selection now takes care of selecting the template.