Limit number of files returned for a process to 50
Computational API allows users to add unlimited numbers of files to a process. As a work around to bringing down the system we will initially limit the number of returned files to 50 and figure out how to best deal with the issue of large numbers of files in a subsequent release.