Fix projects showing up twice
The query to retrieve projects was broken into two parts => user projects, and others. This resulted in users projects showing up twice because projects belonging to the user were not filtered out. A filter is now applied to remove projects belonging to the user. Updated filter to be a part of the query rather than applied on the array afterward.