Show updated username in navbar
When the user updates their username its persisted to the backend. However the local storage where the user information is stored wasn't being updated. Additionally the navbar was never told to update its view. Added updating the local storage and sending a message to the navbar telling it to change the displayed name.