Updates to use session storage
Now that the system is using session storage add calls for properly saving and retrieving state from storage. Right now everything is in one single store that gets rewritten on changes. If this proves to be a bottle neck I'll separate into separate keys in store for individual projects and then decompose further if needed. So far this isn't a problem, and since session storage right now is used for re-establishing state optimizing this performance isn't needed as much.