When getting a project return all the experiment details
Rather than returning just a small amount of details for an experiment when retrieving a project get all the details. Change retrieving files to just getting the file count. Files are a special case because there can be a lot of files but they are rarely accessed. So getting files in a separate call will speed up the system.