Fix highlight for datasets in sidenav when in datasets sub pages
The highlighting wasn't working for sub pages because the initial page was a .list and all other pages were at the same level. Added a service to keep track of the route name and then injected it into the app run to set it on changes, and into the sidenav to check for the datasets route to figure out if the datasets should be highlighted.